+++ src/main/java/com/meta/ponkids/domain/system/ntt/repository/custom/NttReplyRepositoryCustom.java
@@ -1,18 +1,13 @@
 package com.meta.ponkids.domain.system.ntt.repository.custom;
 
-import com.meta.ponkids.domain.system.ntt.dto.NttListDto;
 import com.meta.ponkids.domain.system.ntt.dto.NttReplyListDto;
-
-
 import java.util.List;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
 
 public interface NttReplyRepositoryCustom {
-	List<NttReplyListDto> getList(int nttSn );
-	List<NttReplyListDto> getAnswerReplyList(int nttReplySn );
+	List<NttReplyListDto> getList(Long nttSn );
+	List<NttReplyListDto> getAnswerReplyList(Long nttReplySn );
